% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignerInstallers.R
\name{InstallAligners}
\alias{InstallAligners}
\title{InstallAligners
This function will install and configure the use fo the STAR sofware version 2.7.6a compatible with the Arriba software, and the Rsubread bioconductor package}
\usage{
InstallAligners()
}
\description{
InstallAligners
This function will install and configure the use fo the STAR sofware version 2.7.6a compatible with the Arriba software, and the Rsubread bioconductor package
}
