% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlignerInstallers.R
\name{InstallSTAR}
\alias{InstallSTAR}
\title{InstallSTAR
This function will install and configure the use fo the STAR sofware version 2.7.6a compatible with the Arriba software}
\usage{
InstallSTAR()
}
\description{
InstallSTAR
This function will install and configure the use fo the STAR sofware version 2.7.6a compatible with the Arriba software
}
