% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimgallore.R
\name{InstallTrimGalore}
\alias{InstallTrimGalore}
\title{InstallTrimGalore}
\usage{
\dontrun{
InstallTrimGalore
}
}
\description{
Install [Trimgalore](https://www.bioinformatics.babraham.ac.uk/projects/trim_galore/) priori to check the availability of [cutapat](/cutadapt.readthedocs.io/en/stable/installation.html)
it will install Trimgalore in the Software directory handle by [GenomeDB] library
}
