% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNArunners.R
\name{RunAlign}
\alias{RunAlign}
\title{RunAlign
a wrapper to the \code{\link[Rsubread]{align}} function}
\usage{
RunAlign(sbjFile, species, version, nThreads)
}
\arguments{
\item{sbjFile}{string full path to the R1.fastq or or gz file}

\item{species}{the species you whant to use, it should be stored in your database pls see \code{\link[GenomeDB]{ShowSpecies}}}

\item{version}{the code indicating the genome+annotation pair}

\item{nThreads}{integer, number of CPUs to be used}
}
\description{
RunAlign
a wrapper to the \code{\link[Rsubread]{align}} function
}
