% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runners.R
\name{RunSubjunct}
\alias{RunSubjunct}
\title{RunSubjunct
a wrapper to the \code{\link[Rsubread]{subjuct}} function}
\usage{
RunSubjunct(sbjFile, species, version, nThreads)
}
\arguments{
\item{sbjFile}{string full path to the R1.fastq or or gz file}

\item{species}{the species you whant to use, it should be stored in your database pls see \code{\link[GenomeDB]{ShowSpecies}}}

\item{version}{the code indicating the genome+annotation pair}

\item{nThreads}{integer, number of CPUs to be used}
}
\value{
the file name of the resulted bam file with an attribute "assemblyVersion" that shows the reference-annotation verision 
see \code{\link[GenomeDB]{GetGenomesInDB}}
}
\description{
RunSubjunct
a wrapper to the \code{\link[Rsubread]{subjuct}} function
}
