% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trimgallore.R
\name{RunTrimgalore}
\alias{RunTrimgalore}
\title{RunTrimgalore
RunTrimgalore(fileR1)}
\usage{
\dontrun{
file1 <- "my.file_1.fastq #it is expected to have their pair as ny.file_2.fastq
res <- RunTrimgalore(file1)
res ##this will hold the file1 name and the elapsed time as well as the file sizes prior and after trimming
}
}
\arguments{
\item{fileR1}{the fastq file read 1 (the read 2 will be automatically called) accepted files are XXX_1.fastq/_r1.fastq or fq.gz. Their paired files should replace the _1 (_R1) by _2 (_R2)}

\item{replace}{(boolean) If TRUE the original files will be reomved and replaced by the trimmed ones under the same original name but gziped (.fq.gz), otherwise it will created two new files 
as XXX_1_val_1.fq.gz and XXX_2_val_2,fq.gz}
}
\description{
RunTrimgalores is intended to perform adapters and trimming of read from paired end sequencing data
}
