% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexBuilders.R
\name{UpdateRsubreadindex}
\alias{UpdateRsubreadindex}
\title{UpdateRsubreadindex
It creates the index files for \code{\link[Rsubread]{buildindex}} for each species in the database of \code{\link[GenomeDB]{SetDBdirectory}}}
\usage{
UpdateRsubreadindex()
}
\description{
UpdateRsubreadindex
It creates the index files for \code{\link[Rsubread]{buildindex}} for each species in the database of \code{\link[GenomeDB]{SetDBdirectory}}
}
