% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Runners.R
\name{RunSTAR}
\alias{RunSTAR}
\title{RunSTAR}
\usage{
RunSTAR(sbjFile, species, version, twoPass = c("None", "Basic"), nThreads)
}
\arguments{
\item{sbjFile}{string full path to the R1.fastq or or gz file}

\item{species}{the species you whant to use, it should be stored in your database pls see \code{\link[GenomeDB]{ShowSpecies}}}

\item{version}{the code indicating the genome+annotation pair}

\item{twoPass}{"None" or "Basic"}

\item{nThreads}{integer, number of CPUs to be used}
}
\description{
RunSTAR
}
